#135 [mod] : 상태 관리 방법 변경
