+++ app/src/main/java/com/byeboo/app/presentation/offboarding/offboardingCompletedJourney/OffboardingCompletedJourneyState.kt
@@ -1,7 +1,7 @@
 package com.byeboo.app.presentation.offboarding.offboardingCompletedJourney
 
-import com.byeboo.app.domain.model.offboarding.JourneyType
-import com.byeboo.app.presentation.offboarding.offboardingnewjourney.JourneyStatus
+import com.byeboo.app.presentation.offboarding.model.JourneyCards
+import com.byeboo.app.presentation.offboarding.model.JourneyStatus
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toImmutableList
@@ -9,12 +9,7 @@ import kotlinx.collections.immutable.toImmutableList
 data class OffboardingCompletedJourneyState(
     val journeyCards: ImmutableList<JourneyCards> = persistentListOf()
 ) {
-    val completed: Int = 0
     val completedCards: ImmutableList<JourneyCards>
         get() = journeyCards.filter { it.status == JourneyStatus.COMPLETED }.toImmutableList()
+    val completed: Int get() = completedCards.size
 }
-
-data class JourneyCards(
-    val journeyType: JourneyType,
-    val status: JourneyStatus
-)
